+++ libraries/ui-library-vue/src/lib/stencil-generated/components.ts
@@ -706,6 +706,7 @@ export const SixTab: StencilVueComponent<JSX.SixTab> = /*@__PURE__*/ defineConta
   'panel',
   'active',
   'closable',
+  'hoverContent',
   'disabled',
   'six-tab-close'
 ], [