+++ docs/changelog.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Upcoming
 
+### Added
+
+- Added `hoverContent` prop to `six-tab` to override the default hover content.
+
+### Fixed
+
+### Removed
+
+### Changed
+
+### Deprecated
+
 ## 5.1.0 - 2025-10-23
 
 ### Added